requirements.txt
flask==3.1.0
feedparser==6.0.11
requests==2.32.3
google-generativeai==0.8.3
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.151.0
gunicorn==23.0.0
